requirements.txt
# modulo = version
django
djangorestframework
django-extensions
Pillow